pipeline/requirements.txt
pandas
scikit-learn
joblib
kfp
google-cloud-aiplatform
